Phase 2: Bar accumulators — all 7 variants with overflow rollover
- BaseAccumulator: shared OHLCV tracking, partial bar access (current_bar),
  state persistence (get_state/load_state), bar_id counter
- TickAccumulator: closes on tick count >= threshold
- VolumeAccumulator: closes on cumulative volume >= threshold
- DollarAccumulator: closes on cumulative dollar value >= threshold
- TimeAccumulator: closes on time boundary crossing (clock + first_tick
  anchor); handles empty intervals naturally
- ImbalanceAccumulator: closes on |signed_imbalance| >= threshold;
  tick/volume/dollar metrics; first-tick NaN side excluded from imbalance
- RunAccumulator: closes on cumulative same-sign run statistic >= threshold;
  tick/volume/dollar metrics; first-tick NaN side retroactively included in
  tick-2's run direction per spec; direction-change banks current run

All accumulators implement overflow rollover: excess from the
threshold-crossing tick rolls into the next bar as starting value.

TickInfo.side type changed from int|None to float|None — tick signs are
float64 (derive_tick_sign produces float64; NaN is a float).

62 tests covering: OHLCV tracking, VWAP, bar_id increment, threshold
closure, overflow with hand-computed values, empty accumulator, single-tick
edges, clock/first_tick anchors, multi-bar sequences, first-tick NaN
exclusion/inclusion, signed vs unsigned run accumulation, state round-trip
across all accumulator types, and edge cases (zero volume, float64
precision, threshold=0, large tick counts).

+++ flowbars/core.py
@@ -16,7 +16,7 @@ class TickInfo:
     timestamp: int  # Unix ms (UTC)
     price: float
     volume: float
-    side: int | None = None  # +1 (buy), −1 (sell), or None if not yet derived
+    side: float | None = None  # +1.0 (buy), −1.0 (sell), NaN, or None if not yet derived
 
 
 @dataclass
